backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "langchain[openai]>=0.3.27",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.0.0",
]
